+++ packages/server/migrations/index.js
@@ -4,5 +4,6 @@ import organizations from './0.8.0-organizations';
 import invite from './0.8.3-invite';
 import uniquePath from './0.8.4-unique-path';
 import blockParameters from './0.8.7-block-parameters';
+import nullablePath from './0.8.8-nullable-path';
 
-export default [initial, user, organizations, invite, uniquePath, blockParameters];
+export default [initial, user, organizations, invite, uniquePath, blockParameters, nullablePath];